Solr Upgrade continued  D-3767 D-3768 D-3886
https://marmot.myjetbrains.com/youtrack/issue/D-3767
https://marmot.myjetbrains.com/youtrack/issue/D-3768
https://marmot.myjetbrains.com/youtrack/issue/D-3886

+++ data_dir_setup/update_solr_files.sh
@@ -7,9 +7,12 @@ if [ -z "$1" ]; then
   echo ""
   exit 1
 else
-	rm -f /data/vufind-plus/$1/solr_master/lib/*
-	cp -r solr_master /data/vufind-plus/$1
-	rm -f /data/vufind-plus/$1/solr_searcher/lib/*
-	cp -r solr_searcher /data/vufind-plus/$1
-	exit 0
+	echo "Please make sure existing Solr engines are off."
+	read -p "Proceed with SOLR configuration updates?" -n 1 -r
+	echo    # (optional) move to a new line
+	if [[ $REPLY =~ ^[Yy]$ ]]; then
+		cp -r solr_master /data/vufind-plus/$1
+		cp -r solr_searcher /data/vufind-plus/$1
+		exit 0
+	f1
 fi
